Reworked C++ function
(x) mx functions now take sigma/angle values directly instead of working
with q values as input (initial guesses) and ouput
(x) fixed a critical bug that only the last given initial parameter was
actually taken ..
(x) arbitrary combinations of parameters can be optimized / set
constant. Before the the compuation was reduced to the few relevant
cases for Track'N'Trace, without giving the user an idea of what
happened
(x) Better output if DEBUG is defined + simplified some code

Note: optimizing in q values seemed a bit more instable than in sigma,
as using PSFs of stddev 0.75 with a 0.75 initial guess for the std dev
does not converge every time, in contrast to the old version which
directly worked with the sigmas. (initial guesses smaller than the true
sigma seem to be worse now)